+++ flutter_inappwebview/lib/src/in_app_browser/in_app_browser.dart
@@ -545,4 +545,7 @@ class InAppBrowser implements PlatformInAppBrowserEvents {
 
   @override
   void onMainWindowWillClose() {}
+
+  @override
+  void onProcessFailed(ProcessFailedDetail detail) {}
 }