+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_inappwebview_widget.dart
@@ -127,6 +127,7 @@ class PlatformInAppWebViewWidgetCreationParams
       super.onCameraCaptureStateChanged,
       super.onMicrophoneCaptureStateChanged,
       super.onContentSizeChanged,
+      super.onProcessFailed,
       super.initialUrlRequest,
       super.initialFile,
       super.initialData,